+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.3.0 - 2026-05-
 	- add : weekly ranking
 	- add : ranking for other users
+	- change : no more medal for non drinkers
 	- change : message if no drinker for a ranking
 	- change : syntax modifications