+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2.3.0 - 2026-05-
+## 2.3.0 - 2026-05-22
 	- add : weekly ranking
 	- add : ranking for other users
 	- change : no more medal for non drinkers